requirements.txt
django==4.1.5
django-ckeditor==6.5.1
django-filter==22.1
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
drf-yasg==1.21.5